Toveri interior 3D model v2.
Changed barcode reader mount to post-install, so it is easier to maintain/assemble,
so the barcode reader is attached from the top down.
This way one does not have to disassemble the whole Toveri to change a barcode reader.

Moved the OLED 2mm closer to the window, since the borderline viewing angles lose
visibility to the upper row of the OLED display.

Removing unnecessary material from the pieces, by cutting holes in large flat surfaces.
Managed to make the print cost 5$ lower.